Resolved Issue 8: varying gap dendrogram / heatmap
As pointed out by piatekmj, plotting an annotated heatmap repeatedly on
the same device led to different results, in that there was a small gap
between dendrogram and heatmap the first time, but not any following time.

As expected, this was due to messing with graphical parameters via par()
without taking precautions via on.exit() (specifically xaxs/yaxs). This
has been fixed.