Changed documentation to roxygen
Rd files have been incorporated into the /R source files via Rd2roxygen::Rd2roxygen, with
small manual changes:

* documentation for functions modifyExistingList and extractArg was moved to their
  definition in annHeatmap.R (not recognized by Rd2roxygen)

* Somewhat mysteriously, the \VignetteIndexEntry-command at the top of
  annHeatmapCommentedSource.Rnw was uncommented by Rd2roxygen - this has been
  recommented.

This is the first step in making the package work with the usual devtools-toolchain,
and giving up in the original attempt at some kind of literate programming. This also
means a bump in the version number.